requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cloudinary==1.41.0
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
cssselect2==0.7.0
defusedxml==0.8.0rc2
dj-database-url==2.3.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-rest-swagger==2.2.0
django-restframework==0.0.1
django-templated-mail==1.1.1
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
environ==1.0
fonttools==4.55.3
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
msgpack==1.1.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyphen==0.17.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
simplejson==3.19.3
six==1.17.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
webencodings==0.5.1
whitenoise==6.4.0
zope.interface==7.0.3
zopfli==0.2.3.post1
